+++ packages/web-core/src/services/connection.service.ts
@@ -182,10 +182,10 @@ export class ConnectionService<
     }
   }
 
-  close() {
+  async close() {
     this.cancelScheduler();
     this.state = ConnectionState.INACTIVE;
-    this.connection.close();
+    await this.connection.close();
     this.clearQueue();
   }