+++ packages/web-core/src/services/inworld_connection.service.ts
@@ -90,10 +90,10 @@ export class InworldConnectionService<
   }
 
   async close() {
-    this.connection.close();
+    this.player.clear();
     this.recorder.stop();
     await this.player.stop();
-    this.player.clear();
+    await this.connection.close();
   }
 
   isActive() {